filetools: Rename toolbox.files to file
